Fix Bug, Paginations
I found a problem with page query. If we're in bookings with status unconfirmed and there is for e.g. 30 results, and we go to the third page (with pagination) on the unconfirmed bookings page and go to checked-in page (we filter it) our page will break because page query is 3 and in checked-in page there are only 12 result that mean that we pass the range for checked-in page.
http://localhost:5173/bookings?page=3&status=unconfirmed

http://localhost:5173/bookings?page=3&status=checked-in

To fix that we need to reset the page query every time we change the status

In the Filter.jsx in the handleClick function we add this line of code searchParams.set('page', 1);  before useSearchParams function.

http://localhost:5173/bookings?page=3&status=unconfirmed

http://localhost:5173/bookings?page=1&status=checked-in